compose/transaction-mocker/app/requirements.txt
confluent_kafka
ksql
nltk
kafka-python
faker
confuse
cx_Oracle
